backend/requirements/base.txt
Django>=4.2,<5.0
djangorestframework>=3.14,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-yasg>=1.21,<2.0
django-cors-headers>=4.3,<5.0
django-environ>=0.11,<1.0
django-filter>=23.5,<24.0
django-jazzmin>=2.6,<3.0
psycopg2-binary>=2.9,<3.0
google-auth>=2.25,<3.0
